refactor: recreate the escher builder whenever map changes
There have been multiple reported issues of state not being synchronized
when changing maps. Additionally, there was the "double vision" bug. And
finally, knockouts from visualized design jobs were not always shown
correctly on the map.

This change completely destroys and recreates Escher when map changes
and appears to solve all above issues.

As a consequence, the `onEscherReady` promise is no longer relevant.
Instead, watches will check if the builder exists, and abort if it
doesn't. Then, when Escher is ready, state is always reapplied to make
sure that data for any aborted watcher is applied correctly.