fix: prevent compiler from using editor's fallback $promisedDialog type
Compiler prints type errors in comparison to last type declaration. So
the typesafe declaration needs to be last, but then the fallback
declaration would match and hide type errors. Fixed by separating by
type of `this.value` in component.

If `value` type is void (shims-vue.d.ts): ignore errors
If `value` has type: return that type in a promise and validate props